Fix circular importing
* move functions get_warp_matrix and warp_affine_joints into core
