+++ mmpose/core/post_processing/__init__.py
@@ -1,11 +1,13 @@
 from .nms import oks_nms, soft_oks_nms
 from .post_transforms import (affine_transform, flip_back, fliplr_joints,
                               fliplr_regression, fliplr_regression_3d,
-                              get_affine_transform, rotate_point,
-                              transform_preds)
+                              get_affine_transform, get_warp_matrix,
+                              rotate_point, transform_preds,
+                              warp_affine_joints)
 
 __all__ = [
     'oks_nms', 'soft_oks_nms', 'affine_transform', 'rotate_point', 'flip_back',
     'fliplr_joints', 'fliplr_regression', 'fliplr_regression_3d',
-    'transform_preds', 'get_affine_transform'
+    'transform_preds', 'get_affine_transform', 'get_warp_matrix',
+    'warp_affine_joints'
 ]
